dom/ipc/tests/test_Preallocated.html

<!DOCTYPE HTML>
<html>
<!--
Test that the preallocated process starts up.
-->
<head>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="../browserElementTestHelpers.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>

<script type="application/javascript">
"use strict";

SimpleTest.waitForExplicitFinish();

function expectProcessCreated() {
  /* eslint-env mozilla/chrome-script */
  return new Promise(resolve => {
    function parentExpectProcessCreated() {
      let topic = "ipc:content-initializing";
      let obs = { observe() {
        Services.obs.removeObserver(obs, topic);
        sendAsyncMessage("process-created");
      }};
      Services.obs.addObserver(obs, topic);
    }

    let helper = SpecialPowers.loadChromeScript(parentExpectProcessCreated);
    SimpleTest.registerCleanupFunction(function() { helper.destroy(); });
    helper.addMessageListener("process-created", resolve);
  });
}

expectProcessCreated().then(() => {
  ok(true, "Process creation detected.");
  SimpleTest.finish();
});

// Kill existing preallocated process.
SpecialPowers.pushPrefEnv({"set": [["dom.ipc.processPrelaunch.enabled", false]]}).then(() => {
  // Make sure we have the capacity to launch preallocated process.
  SpecialPowers.pushPrefEnv({"set": [["dom.ipc.processCount", 100]]}).then(() => {
    // Enable preallocated process and run the test.
    SpecialPowers.pushPrefEnv({"set": [["dom.ipc.processPrelaunch.enabled", true]]});
  });
});
</script>
</body>
</html>
